Refactor: Delete old terminal scripts and professionally organize backend src/ directory
