+++ src/core/analysis.py
