+++ src/services/stock_service.py
